requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
hdbscan>=0.8.33
umap-learn>=0.5.5
plotly>=5.18.0
ipykernel>=6.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.13.0
tenacity>=8.2.3
